Create Polymorphic association for Comments. Work on Tube Polymorphism

+++ ampsite/db/migrate/20150617184833_pentodes.rb
@@ -3,13 +3,13 @@ class Pentodes < ActiveRecord::Migration
     create_table :pentodes do |t|
       t.string :identifier, null: false
       t.string :classification, null: false
-      t.integer :filament_voltage, null: false
-      t.integer :filament_current, null: false
-      t.integer :max_anode_voltage, null: false
-      t.integer :anode_dissipation, null: false
-      t.integer :max_grid2_voltage, null: false
-      t.integer :max_grid2_dissipation, null: false
-      t.integer :amplification_factor, null: false
+      t.float :filament_voltage, null: false
+      t.float :filament_current, null: false
+      t.float :max_anode_voltage, null: false
+      t.float :anode_dissipation, null: false
+      t.float :max_grid2_voltage, null: false
+      t.float :max_grid2_dissipation, null: false
+      t.float :amplification_factor, null: false
 
       t.timestamps
     end
